+++ src/test/java/org/geowe/service/geometry/test/SymDifferenceTest.java
@@ -118,4 +118,16 @@ public class SymDifferenceTest {
 		Assert.isTrue(differenceGeometries.size() == 13);
 	}
 
+	@Test
+	public void symDifferenceLinesFeatureCollectionElements(){
+		OperationData opData = DataTestProvider.getLinesFCIntersectionData();
+		target = restClient.target(SERVICE_URL+"/elements");
+		Response response = target.queryParam("tolerance", 0.0001).request().post(
+				Entity.entity(opData,"application/json;charset=UTF-8"));
+
+		List<String> differenceGeometries = response.readEntity(new GenericType<List<String>>(){});
+		response.close();
+		Assert.isTrue(response.getStatus() == Status.CREATED.getStatusCode());
+		Assert.isTrue(differenceGeometries.size() == 5);
+	}
 }